+++ changelog.md
@@ -2,6 +2,8 @@
 
 ## Upcoming
 ### Changes
+* Enchanting tables can now upgrade enchantments with same level books
+  * So if the enchanting table knows Efficiency III, an Efficiency III book will unlock IV
 * Repair Kits now require half the smashes to be crafted (~~10~~ -> 5)
 * Increased max crates before slowdown to 2
 * Added missing Caverns and Chasms Repair kits